pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "purkinje-uv"
version = "0.2.2"
description = "Utilities to generate Purkinje networks from fractal trees and mesh surfaces"
authors = [{ name = "Ricardo García Ramírez", email = "rgr.5882@gmail.com" }]
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
dependencies = [
    "numpy",
    "scipy",
    "vtk",
    "pyvista",
    "meshio",
    "fim-python",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[project.urls]
"Homepage" = "https://github.com/ricardogr07/purkinje-uv/"
"Repository" = "https://github.com/ricardogr07/purkinje-uv/"

[tool.black]
line-length = 88
target-version = ["py310"]
skip-string-normalization = true

[tool.ruff]
target-version = "py310"

[tool.mypy]
files = ["src"]
python_version = "3.10"
strict = true
ignore_missing_imports = true
disallow_untyped_calls = false
check_untyped_defs = true

[tool.pydocstyle]
convention = "google"
match      = ".*\\.py"
add-ignore = ["D104", "D105", "D107"]

[tool.pytest.ini_options]
pythonpath = ["src"]
addopts = "--cov=purkinje_uv --cov-fail-under=80 -ra -q"
testpaths  = ["tests"]

[tool.tox]
envlist                    = ["py310", "py312", "lint"]
isolated_build             = true
skip_missing_interpreters  = true

[tool.tox.testenv]
description = "Run the test suite"
deps = [
  "pytest>=7.1,<8.0",
]
commands = ["pytest tests"]

[tool.tox.envs.lint]
description = "Run formatting, static analysis, and style checks"
skip_install  = true
deps          = ["black", "ruff", "mypy", "pydocstyle"]
commands      = [
  "black --check src/ tests/",
  "ruff check src/ tests/",
  "mypy src/",
]